SILA-4085 - Request changes have been done and structured generate processor components.
