+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.6.3",
     "bootstrap": "^4.6.0",
+    "dompurify": "^2.3.6",
     "js-cookie": "^3.0.1",
     "react": "^17.0.1",
     "react-bootstrap": "^1.4.3",
